Moving stan/services/init/init_*.hpp to stan/services/sample/
