--- src/test/unit/services/init/init_static_hmc_test.cpp
@@ -1,3 +1,0 @@
-#include <stan/services/init/init_static_hmc.hpp>
-#include <gtest/gtest.h>
-// FIXME: needs tests
