--- src/test/unit/services/init/init_windowed_adapt_test.cpp
@@ -1,3 +1,0 @@
-#include <stan/services/init/init_windowed_adapt.hpp>
-#include <gtest/gtest.h>
-// FIXME: needs tests
