+++ src/test/unit/services/sample/init_adapt_test.cpp
@@ -1,3 +1,3 @@
-#include <stan/services/init/init_adapt.hpp>
+#include <stan/services/sample/init_adapt.hpp>
 #include <gtest/gtest.h>
 // FIXME: needs tests
