+++ src/test/unit/services/sample/init_nuts_test.cpp
@@ -1,3 +1,3 @@
-#include <stan/services/init/init_nuts.hpp>
+#include <stan/services/sample/init_nuts.hpp>
 #include <gtest/gtest.h>
 // FIXME: needs tests
