+++ src/test/unit/services/sample/init_static_hmc_test.cpp
@@ -1,0 +1,3 @@
+#include <stan/services/sample/init_static_hmc.hpp>
+#include <gtest/gtest.h>
+// FIXME: needs tests
