+++ src/test/unit/services/sample/init_windowed_adapt_test.cpp
@@ -1,0 +1,3 @@
+#include <stan/services/sample/init_windowed_adapt.hpp>
+#include <gtest/gtest.h>
+// FIXME: needs tests
